# Conflicts:
#	src/main/java/com/example/idatt2106_2022_05_backend/controller/UserController.java
#	src/main/java/com/example/idatt2106_2022_05_backend/dto/ad/AdDto.java
#	src/main/java/com/example/idatt2106_2022_05_backend/dto/user/UserUpdateDto.java
#	src/main/java/com/example/idatt2106_2022_05_backend/model/User.java
#	src/main/java/com/example/idatt2106_2022_05_backend/service/ad/AdService.java
#	src/main/java/com/example/idatt2106_2022_05_backend/service/ad/AdServiceImpl.java
#	src/main/java/com/example/idatt2106_2022_05_backend/service/authorization/AuthServiceImpl.java
#	src/main/java/com/example/idatt2106_2022_05_backend/service/review/ReviewServiceImpl.java
#	src/main/java/com/example/idatt2106_2022_05_backend/service/user/UserDetailsServiceImpl.java
#	src/main/resources/application.yml

+++ src/main/java/com/example/idatt2106_2022_05_backend/service/review/ReviewServiceImpl.java
@@ -50,7 +50,6 @@ public class ReviewServiceImpl implements ReviewService{
         }
     }
 
-
     /**
      * Method to validate that a user only posts once per ad
      * @param ad ad
@@ -77,19 +76,46 @@ public class ReviewServiceImpl implements ReviewService{
 
         review.setDescription(newReviewDto.getDescription());
         review.setRating(newReviewDto.getRating());
+
         //checking that the same user does not post twice per ad
-        User user = getUser(newReviewDto.getUserId());
-        Ad ad = getAd(newReviewDto.getAdId());
+        User user = userRepository.getById(newReviewDto.getUserId());
+        Optional<Ad> ad = adRepository.findById(newReviewDto.getAdId());
+        if(ad.isPresent()){
+            if(validateUser(ad.get(), user)) {
 
-        if(!validateUser(ad, user)){
-            return new Response("en bruker kan kun post 1 omtale per annonse", HttpStatus.BAD_REQUEST);
+                review.setUser(user);
+
+                // Setting ad
+                review.setAd(ad.get());
+
+                Review reviewSaved = reviewRepository.save(review);
+
+                // Set the review to the list of reviews for the user
+                user.getReviews().add(reviewSaved);
+                //user.addReview(reviewSaved);
+
+                // Persist the change
+                userRepository.save(user);
+
+                // Retrieve the user that owns the ad
+                User ownerOfAd = ad.get().getUser();
+
+                // Increment the number of reviews for the user
+                ownerOfAd.setNumberOfReviews(user.getNumberOfReviews() + 1); // todo if getNumberOfReviews == null implement check
+
+                // Add the rating to the total rating of the user
+                ownerOfAd.setRating(user.getRating() + newReviewDto.getRating());
+
+                // Persist the users changes
+                userRepository.save(ownerOfAd);
+
+                return new Response("Omtalen ble lagret", HttpStatus.OK);
+            }
+            else {
+                return new Response("en bruker kan kun post 1 omtale per annonse", HttpStatus.BAD_REQUEST);
+            }
         }
-        review.setUser(user);
-        //Setting ad
-        review.setAd(ad);
-
-        reviewRepository.save(review);
-        return new Response("Omtalen ble lagret", HttpStatus.OK);
+        return new Response("Kunne ikke finne en annonse med gitt ID.", HttpStatus.BAD_REQUEST);
     }
 
 
@@ -102,51 +128,74 @@ public class ReviewServiceImpl implements ReviewService{
      */
     @Override
     public Response getReviewsByAdId(long ad_id){
-        Ad ad = getAd(ad_id);
-        Set<ReviewDto> reviews = reviewRepository.getAllByAd(ad).stream()
-                .map(review -> modelMapper.map(review, ReviewDto.class)).collect(Collectors.toSet());
-        //Returns reviews if found
-        if(reviews.size()!=0) {
+        Optional<Ad> ad = adRepository.findById(ad_id);
+        if(ad.isPresent()) {
+            List<ReviewDto> reviews = reviewRepository.getAllByAd(ad.get()).stream()
+                    .map(review -> modelMapper.map(review, ReviewDto.class)).collect(Collectors.toList());
+
+            // Returns reviews
             return new Response(reviews, HttpStatus.OK);
         }
-        return new Response("fant ingen omtaler på denne annonsen", HttpStatus.NOT_FOUND);
+        else {
+            return new Response("fant ingen omtaler på denne annonsen", HttpStatus.NOT_FOUND);
+        }
     }
 
     /**
-     * Method to delete a review.
-     * @param ad_id id of ad to be deleted.
-     * @param user_id user who wrote the review.
-     * @return response.
+     * method to delete a review
+     *
+     * @param ad_id id of ad to be deleted
+     * @param user_id user who wrote the review
+     *
+     * @return response
      */
     @Override
     public Response deleteReview(long ad_id, long user_id){
-        Optional<Review> review = reviewRepository.getByAdAndUser(adRepository.getById(ad_id), userRepository.getById(user_id));
+        Optional<Review> review = reviewRepository.getByAdAndUser(adRepository.getById(ad_id),
+                             userRepository.getById(user_id));
         if(review.isPresent()){
-            reviewRepository.delete(review.get());
+
+            // Remove the rating from the original ad and user
+            Optional<User> userFound = userRepository.findById(review.get().getUser().getId());
+            Optional<Ad> adFound = adRepository.findById(ad_id);
+            if(userFound.isPresent() && adFound.isPresent()) {
+
+                // If the user who wrote the review is the user trying to delete it
+                if(userFound.get().getId().equals(user_id)) {
+
+                    // Remove the review from the user that created it
+                    userFound.get().getReviews().remove(review.get());
+                    adFound.get().getReviews().remove(review.get());
+
+                    // Find the user that owns the ad
+                    User adOwner = adFound.get().getUser();
+
+                    // Remove the rating from that users total rating
+                    adOwner.getReviews().remove(review.get());
+                    adOwner.setRating(adOwner.getRating() - review.get().getRating());
+
+                    // Decrement that users total number of ratings
+                    adOwner.setNumberOfReviews(adOwner.getNumberOfReviews() - 1);
+
+                    // Persist the user that owns the ad
+                    userRepository.save(adOwner);
+
+                    // Persist the user that deleted the ad and the ad
+                    userRepository.save(userFound.get());
+                    adRepository.save(adFound.get());
+
+                    // Remove the foreign keys from the review
+                    review.get().setUser(null);
+                    review.get().setAd(null);
+
+                    // Delete the review
+                    reviewRepository.delete(review.get());
+                }
+            }
             return new Response("Omtalen ble slettet", HttpStatus.OK);
         }
-        return new Response("fant ikke omtalen", HttpStatus.NOT_FOUND);
-    }
-
-    //Private support methods:
-
-    /**
-     * Helper method to retrieve user by id.
-     * @param id id of user.
-     * @return user found.
-     */
-    private User getUser(long id){
-        return userRepository.findById(id)
-                .orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND, "Fant ikke brukeren"));
-    }
-
-    /**
-     * Helper method to retrieve Ad by id.
-     * @param id id of ad
-     * @return ad found.
-     */
-    private Ad getAd(long id){
-        return adRepository.findById(id)
-                .orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND, "fant ikke annonsen"));
+        else {
+            return new Response("fant ikke omtalen", HttpStatus.NOT_FOUND);
+        }
     }
 }
